pca-web/backend/requirements.txt
fastapi
uvicorn
python-multipart
fastapi
uvicorn
python-multipart
numpy
pillow
scikit-image
opencv-python-headless
scipy
matplotlib
